+++ examples/questions.js
@@ -1,7 +1,7 @@
-import { Questions } from "interactivizer";
+import { Questions, ask } from "interactivizer";
 
 // Create a questions class
-const questions = new Questions(["First question?", "Second question?"]);
+const questions = new Questions({questions: ["First question?", "Second question?"]});
 
 // Or create an empty question and add them later
 const questions2 = new Questions();
@@ -17,3 +17,6 @@ questions2.addQuestion("First question?").addQuestion("Second question?");
 //   } 
 // ]
 const answers = await questions.ask();
+
+// Ask a single question
+ask("Some question?");