+++ CHANGELOG.md
@@ -16,3 +16,15 @@ This release fixes a bug in v0.1.0.
 
 ### Fixed
 - Logic of `Questions` constructor's parameter.
+
+## v0.1.2 - (2022-12-30)
+
+Added some more functionality, and some minor changes.
+
+### Added
+- `ask()` that acts like `Questions` but for only one question.
+- Example for `ask()`
+
+### Changed
+- `Questions` constructor parameters are changed to options.
+- Logic handling parameters for `Questions`